SDK: Ensure that AWS_SDK_CONFIG_OPT_OUT is set
Amazon added the ability to assume roles through the SDK using profiles
defined in the default ~/.aws/config a while ago. However, this
behaviour interferes with aws-runas's workflow when a local default
config is defined without any corresponding credentials. I'm personally
using this pattern on AWS instances that do not need any credentials
defined (ie: instance profiles) where it would make more sense to use
~/.aws/config versus a local aws_config.

Since half of the reason this tool was created in the first place was to
work around the lack of this kind of behaviour in the SDK ecosystem in
general (not just on Ruby), this extra functionality just gets in the
way. This commit includes a test to ensure that we can assume a role
given a barebones config file with only a role_arn defined (and a
region, but that doesn't come into play too much here), and of course
the env variable, which is set before the gem is loaded in
AwsRunAs::Main.