Ensure that AWS_REGION does not get set unless we have a region
This ensures that we only set AWS_REGION if we have AWS_REGION in
config. Before this, it setting it regardless, even if the return value
was nil.

The side effect before was that AWS_REGION was never loaded as an
environment variable, as it doesn't look like nil values make it into
the environment of a launched process/shell. However, this can't be
tested on this way, as we are testing what the raw hash is set to,
which includes the value set as nil.