zsh prompt support, bash/zsh extra feedback functionality
zsh support has now been added for the colorized prompt.

Also, for both bash and zsh, a Bourne-compatible shell profile is now
inserted with two functions - session_expired, which will check to see
if a session has expired based off of the AWS_SESSION_EXPIRATION_UNIX
variable, and session_status_color, which renders a color depending on
whether or not the session has expired. These are used by the prompt
colorization process but can also be utilized externally.