Add --skip-prompt to CLI arguments
This allows someone to skip the fancy bash prompt, but keep the helper
functions that it relies on, and the environment that it works off of.

This logic is generally handled in Ruby code still (including a static
setting of the profile name), but the groundwork is pretty much laid now
for future refactorings (if necessary) to run most of this logic now in
shell scripts thrown in shell_profiles. The sh.profile script in there
now could handle prompt logic for both bash and zsh using conditionals.
For now though, it's fine the way it is.