Add duration_seconds to get_session_token
You can set anywhere up to a maximum of 3 days for the maximum on
get_session_token, so you will be able to do the full 12 hours you would
be able to do on AssumeRole. According to the AWS SDK, these values are
truncated if they are out of bounds.