Refactor session ID to use API
This refactors the session ID logic in this PR to use GetCallerIdentity
from the STS API.

This presents a more secure method of ensuring that the user name is
embedded in the session ID - relying on client environment for this is
easily defeatable simply by passing a different USER environment
variable to aws-runas.

If GetCallerIdentity is not available due to access issues, we just
resort the old behaviour of just using a slug and the unix timestamp.

It also exposes the session ID to the environment.