Main: Correct some session length/nesting issues
* When calling aws-runas from an assumed role, the session name now
takes on the access key ID of the user in GetCallerIdentity, not the
account ID and session name. This helps prevent nesting issues where the
aws-runas-session_BLAH_TIMESTAMP session name would repeat, causing
session name length issues, aside from just being confusing.
* We now check the generated session name overall, and if it's over 64
characters, we revert back to the classic aws-runs-session_TIMESTAMP
syntax. This may be modified in the future to use the access key ID
instead in the event that it's due to account/user length.

Fixes #17.