Main: More session ID generation stuff
Updated the code to add another fallback to an access key ID-based
session name if using the account ID and user would end up rendering a
session ID longer than 64 characters (such as in the event of long user
names).

The classic session ID is now purely a fallback for edge cases where
both ID schemes are completely impossible, or if the user has no access
to GetCallerIdentity.